colors
team, i've changed the appbar in 5 places... all different colors. so that as we use it, see if any color resonates with you and the theme... we can always go back to brown[900] which we are comfortable with... if you have thoughts of colors to try, also let me know and we can add on. the appbar colors are updated here 1. main 2. detail 3. search 4. personalize 5. category